# dist/api.yml
openapi: 3.0.0
info:
  title: api
  description: API for  project
  version: 1.0.0
...
paths:
...
  /unionAgent/leagueManagementPage:
    get:
      tags:
        - UnionAgent
      summary: Return all league matches
      description: All league matches that exists in the system
      parameters: 
      - in: query
        name: sortBy
        description: Sort matches by "Date" or team name or leave empty 
        schema:
          type: string
          example: "Date"
      responses:
        '200':
          description: The league management page uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  pastMatches:
                    type: array
                    items:
                      $ref: '#/components/schemas/PastMatchInformation'
                  futureMatches:
                    type: array
                    items:
                      $ref: '#/components/schemas/FutureMatchInformation'
        "204":
          description: No Content
        '400':
          description: Bad request
  /unionAgent/match:
    get:
      tags:
        - UnionAgent
      summary: Return all the data the agent need for new match
      responses:
        "200":
          description: Data the agent need for new match returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  teamsNames:
                    type: array
                    items:
                      type: string
                      description: Team name in our league
                      example: "Midtjylland"
                  venuesNames:
                    type: array
                    items:
                      type: string
                      description: Venue name in out league
                      example: "MCH Arena"
                  referees:
                    type: array
                    items:
                     $ref: '#/components/schemas/Referee'
        "401":
          description: User is not authorized
    post:
      tags: 
        - UnionAgent
      summary: Add match to the league
      description: Add match to the league by union agent user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                matchInformation:
                  type: object
                  properties:
                    matchDate:
                      type: string
                      format: date-time
                      description: Match's date and time of beginning
                      example: "2021-07-02 12:00:00"
                    localTeamName:
                      type: string
                      description: Local team Name
                      example: Viborg
                    visitorTeamName:
                      type: string
                      description: Visitor team name
                      example: Silkeborg
                    venueName:
                      type: string
                      description: Venue name
                      example: Sydbank Park
                refereeID:
                  type: number
                  description: The referee's ID
                  example: 1
      responses:
        '201':
          description: Match added to league's matches successfully
        '400':
          description: Bad request
    delete:
      tags: 
        - UnionAgent
      summary: Delete match from the league
      description: Delete match from the league by union agent user
      parameters:
      - in: query
        name: matchID
        description:  Match's id needed to delete
        schema:
          type: integer
          example: 20
      responses:
        '200':
          description: Match deleted from league's matches successfully
        '400':
          description: Bad request
        '404':
          description: Match not found
  /unionAgent/addMatchResult:
    get:
      tags:
        - UnionAgent
      summary: Return all the data the agent need for add match result
      responses:
        "200":
          description: Data the agent need for add match result
          content:
            application/json:
              schema:
                type: object
                properties:
                  pastMatchesWithoutResult:
                    type: array
                    items:
                      $ref: '#/components/schemas/PastMatchInformation'
        "204":
          description: No Content
        "401":
          description: User is not authorized
    put:
      tags: 
        - UnionAgent
      summary: Add result to match
      description: Add result to match in the league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matchID:
                  type: number
                  description: The match's ID to add result
                localTeamScore:
                  type: number
                  description: The number of goals scored by the home team
                visitorTeamScore:
                  type: number
                  description: The number of goals scored by the home team
              example: 
                matchID: 939
                localTeamScore: 5
                visitorTeamScore: 1
      responses:
        '200':
          description: Result added to match successfully
        '400':
          description: Bad request
  /unionAgent/addMatchEventsLog:
    get:
      tags:
        - UnionAgent
      summary: Return all the data the agent need for add match event log
      responses:
        "200":
          description: Data the agent need for add match match event log
          content:
            application/json:
              schema:
                type: object
                properties:
                  pastMatchesWithAllInfo:
                    type: array
                    items:
                      $ref: '#/components/schemas/PastMatchInformation'
        "204":
          description: No Content
        "401":
          description: User is not authorized
    post:
      tags: 
        - UnionAgent
      summary: Add events log to match
      description: Add events log to match in the league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matchID:
                  type: number
                  description: The match's ID to add events log
                  example: 939
                eventsLog:
                  type: array
                  description: Match's event log
                  items:
                    $ref: '#/components/schemas/MatchEventInsertion'
      responses:
        '200':
          description: Events log added match successfully
        '400':
          description: Bad request
    delete:
      tags: 
        - UnionAgent
      summary: Delete match event 
      description: Delete event from match's events log
      parameters:
      - in: query
        name: matchID
        description:  The match's ID
        schema:
          type: integer
          example: 20
      - in: query
        name: eventID
        description:  Event's ID to delete
        schema:
          type: integer
          example: 10
      responses:
        '200':
          description: Event deleted from match's events log successfully
        '400':
          description: Bad request
  /unionAgent/addRefereeToMatch:
    get:
      tags:
        - UnionAgent
      summary: Return all referees and matches without referee
      responses:
        "200":
          description: Data the agent need for add referee returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  referees:
                    type: array
                    items:
                      $ref: '#/components/schemas/Referee'
                  futureMatchesWithoutReferee:
                    type: array
                    items:
                      $ref: '#/components/schemas/FutureMatchWithoutReferee'
                  pastMatchesWithoutReferee:
                    type: array
                    items:
                      $ref: '#/components/schemas/PastMatchWithoutReferee'
        "401":
          description: User is not authorized
    post:
      tags: 
        - UnionAgent
      summary: Add referee to match
      description: Add referee to match in the league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matchID:
                  type: number
                  description: The match's ID to add referee
                  example: 939
                refereeID:
                  type: number
                  description: The referee's ID
                  example: 252
      responses:
        '200':
          description: Referee added to match successfully
        '400':
          description: Bad request
    put:
      tags: 
        - UnionAgent
      summary: Update referee in match
      description: Update referee in match in the league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matchID:
                  type: number
                  description: The match's ID to add referee
                  example: 939
                refereeID:
                  type: number
                  description: The referee's ID
                  example: 252
      responses:
        '200':
          description: Referee updated in match successfully
        '400':
          description: Bad request
...
components:
  schemas:
...
    FutureMatchInformation:
      type: object
      description: The information about future match
      required:
        - matchDate
        - localTeamName
        - visitorTeamName
        - venueName
      properties:
        matchID:
          type: number
          description: Match's ID
          example: 1 
        matchDate:
          type: string
          format: date-time
          description: Match's date and time of beginning
          example: "2021-07-02 12:00:00"
        localTeamName:
          type: string
          description: The local team name
          example: Viborg
        visitorTeamName:
          type: string
          description: The visitor team name
          example: Silkeborg
        venueName:
          type: string
          description: The match stadium name
          example: Sydbank Park
        refereeInformation:
          $ref: '#/components/schemas/Referee'
    PastMatchInformation:
      type: object
      description: The information about past match
      required:
        - matchDateAndTime
        - localTeamName
        - visitorTeamName
        - venueName
        - localTeamScore
        - visitorTeamScore
        - eventsLog
      properties:
        matchID:
          type: number
          description: Match's ID
          example: 1 
        matchDateAndTime:
          type: string
          format: date-time
          description: Match's date and time of beginning
          example: "2020-08-02 15:30:00"
        localTeamName:
          type: string
          description: The local team name
          example: Midtjylland
        visitorTeamName:
          type: string
          description: The visitor team name
          example: AaB
        venueName:
          type: string
          description: The match field's name
          example: MCH Arena 
        localTeamScore:
          type: number
          description: The number of goals scored by the home team
          example: 5
        visitorTeamScore:
          type: number
          description: The number of goals scored by the home team
          example: 1
        refereeInformation:
          $ref: '#/components/schemas/Referee'
        eventsLog:
          type: array
          description: Match's event log
          items:
            $ref: '#/components/schemas/MatchEvent'
    FutureMatchWithoutReferee:
      type: object
      description: The information about future match
      required:
        - matchDate
        - localTeamName
        - visitorTeamName
        - venueName
      properties:
        matchID:
          type: number
          description: Match's ID
          example: 1 
        matchDate:
          type: string
          format: date-time
          description: Match's date and time of beginning
          example: "2021-07-02 12:00:00"
        localTeamName:
          type: string
          description: The local team name
          example: Viborg
        visitorTeamName:
          type: string
          description: The visitor team name
          example: Silkeborg
        venueName:
          type: string
          description: The match stadium name
          example: Sydbank Park
    PastMatchWithoutReferee:
      type: object
      description: The information about past match
      required:
        - matchDateAndTime
        - localTeamName
        - visitorTeamName
        - venueName
        - localTeamScore
        - visitorTeamScore
        - eventsLog
      properties:
        matchID:
          type: number
          description: Match's ID
          example: 1 
        matchDateAndTime:
          type: string
          format: date-time
          description: Match's date and time of beginning
          example: "2020-08-02 15:30:00"
        localTeamName:
          type: string
          description: The local team name
          example: Midtjylland
        visitorTeamName:
          type: string
          description: The visitor team name
          example: AaB
        venueName:
          type: string
          description: The match field's name
          example: MCH Arena 
        localTeamScore:
          type: number
          description: The number of goals scored by the home team
          example: 5
        visitorTeamScore:
          type: number
          description: The number of goals scored by the home team
          example: 1
        eventsLog:
          type: array
          description: Match's event log
          items:
            $ref: '#/components/schemas/MatchEvent'
    MatchEvent:
      type: object
      description: Match Event
      required:
        - eventTimeAndDate
        - minuteInMatch
        - eventType
      properties:
        eventID:
          type: number
          description: event's ID number
          example: 10
        eventTimeAndDate:
          type: string
          format: date-time
          description: The date and time the event occurred 
          example: "2020-08-02 15:30:00"
        minuteInMatch:
          type: number
          description: The minute in match the event occurred
          example: 10
        eventType:
          type: string
          description: Type of the match event
          example: Goal
          enum: 
          - Goal
          - Offside
          - Red Card
          - Yellow Card
          - Injury
          - Subsitute
        eventDescription:
          type: string
          description: More information about the event
          example: Yellow card to Messi
    MatchEventInsertion:
      type: object
      description: Match Event Insertion Template
      required:
        - minuteInMatch
        - eventType
      properties:
        minuteInMatch:
          type: number
          description: The minute in match the event occurred
          example: 10
        eventType:
          type: string
          description: Type of the match event
          example: Goal
          enum: 
          - Goal
          - Offside
          - Red Card
          - Yellow Card
          - Injury
          - Subsitute
        eventDescription:
          type: string
          description: More information about the event
          example: Yellow card to Messi
    Referee:
      type: object
      description: Referee information
      properties:
        refereeID:
          type: number
          description: Referee's ID
          example: 10
        firstname:
          type: string
          description: Referee's first name
          example: Stefani
        lastname:
          type: string
          description: Referee's last name
          example: Germanotta
        course:
          type: string
          description: Referee's course
          example: Regular
          enum: 
          - Regular
          - Main